test: extract extension from path
