+++ src/main/java/http/HttpResponse.java
@@ -2,17 +2,21 @@ package http;
 
 public class HttpResponse {
     private final int statusCode;
-    private final String contentType;
+    private final ContentType contentType;
     private final byte[] responseBody;
 
-    public HttpResponse(final int statusCode, final String contentType, final byte[] responseBody) {
+    private HttpResponse(final int statusCode, final ContentType contentType, final byte[] responseBody) {
         this.statusCode = statusCode;
         this.contentType = contentType;
         this.responseBody = responseBody;
     }
 
-    public static HttpResponse ok(final String contentType, final byte[] responseBody) {
+    public static HttpResponse ok(final ContentType contentType, final byte[] responseBody) {
         return new HttpResponse(200, contentType, responseBody);
+    }
+
+    public ContentType getContentType() {
+        return contentType;
     }
 
     public int getStatusCode() {
