+++ src/main/java/http/Path.java
@@ -5,6 +5,7 @@ import utils.Tokens;
 public class Path {
     private static final String PATH_AND_QUERY_STRING_DELIMITER = "\\?";
     private static final int MINIMUM_REQUIRED_TOKEN_SIZE = 1;
+    public static final char DOT = '.';
 
     private final String path;
     private final QueryString queryString;
@@ -19,6 +20,16 @@ public class Path {
 
     public String getPath() {
         return path;
+    }
+
+    public String getExtension() {
+        int lastIndexOfDot = path.lastIndexOf(DOT);
+
+        if (lastIndexOfDot < 0) {
+            return null;
+        }
+
+        return path.substring(lastIndexOfDot + 1);
     }
 
     public String getParameter(final String parameterName) {
