+++ src/test/java/http/PathTest.java
@@ -2,6 +2,9 @@ package http;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
+import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
@@ -23,5 +26,14 @@ public class PathTest {
         assertThat(path.getPath()).isEqualTo("/users");
         assertThat(path.getParameter("name1")).isEqualTo("value1");
         assertThat(path.getParameter("name2")).isEqualTo("value2");
+    }
+
+    @ParameterizedTest
+    @CsvSource(value = {"/index.html,html", "/style.css,css", "/test.img,img", "/test.test.js,js"}, delimiter = ',')
+    @DisplayName("path내부에서 확장자가 존재한다면 잘 추출하는지")
+    void getExtension(final String pathStr, final String expected) {
+        Path path = new Path(pathStr);
+
+        assertThat(path.getExtension()).isEqualTo(expected);
     }
 }
